requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
black==23.7.0
click==8.1.5
fastapi==0.100.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.2.0
pydantic==1.*
pypika-tortoise==0.1.6
pytest==7.4.0
pytz==2023.3
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.19.3
typing_extensions==4.7.1
uvicorn==0.23.0
